\name{build.q.set}
\alias{build.q.set}
\title{Q methodology: build q set from sample and concourse}
\description{
	Subsets a concourse of items with a sample of selected items.
	Returns a dataframe with short handles as rownames, and languages as columns.
}
\usage{
	build.q.set(q.concourse, q.sample, q.distribution)
}

\arguments{
  \item{q.concourse}{
  	A matrix with meaningful short handles as rownames, languages as columns, and full items in cells as produced by \code{\link{import.q.concourse}}.
  }
  \item{q.sample}{
  	A character vector of sampled meaningful short handles, such as \code{c("q-is-great","q-is-obscure")}.
  }
  \item{q.distribution}{
  	The chosen q distribution as a vector of integers, such as \code{c(1,3,1)}.
  }
}
\details{
	Q studies are carried out letting participants rank a \emph{sample} of statements (items), collectively referred to as the \emph{q-set}.
	These q-sets are drawn (by some sampling strategy) from  \emph{concourse}, or universe of items.
	This function currently simply subsets the concourse generated by \code{\link{import.q.concourse}} and returns it as \code{q.set}.
	Throws error if the number of sampled items does not correspond to the integral under the defined q-distribution.
}

\value{
	Returns a matrix with short handles as rownames, languages as columnames and full translations in cells.
}

\note{
	This function currently does \emph{not} actually \emph{draw} a sample, but merely builds the q-set from a given sample.
	Comment and/or chip in if you like want an actual sampling feature: \url{https://github.com/aiorazabala/qmethod/issues/5}.

	This function currently requires a q.distribution, but it only checks for the sum, so if you are working with a distribution-free study that still has a fixed number of items (as most will), you can just enter a vector of length 1 with your sum as the q.distribution.
}

\author{Maximilian Held}

\seealso{
	\code{\link{import.q.concourse}},
	\code{\link{import.q.feedback}},
	\code{\link{import.q.sorts}},
	\code{\link{make.cards}}
}