\name{keyneson}
\docType{data}
\alias{keyneson}
\title{Keyneson Q sort on taxation and the economy}
\description{A Q-methodological study on taxation and the economy, including a concourse, sample and Q-sorts}
\usage{keyneson}
\format{A directory in the package root, including subdirectories as documented in the package documentation, and on the package \href{https://github.com/aiorazabala/qmethod/wiki/Data-management}{wiki}.}
\source{Maximilian Held. (2014). keyneson: Pretest (with friends, students). ZENODO. 10.5281/zenodo.12678}
\keyword{datasets}
