\name{lipset}
\docType{data}
\alias{lipset}
\title{\emph{Lipset} (1963) Q methodology dataset}
\description{Lipset (1963) dataset about \emph{The Value Patterns of Democracy} to illustrate the \pkg{qmethod} package.}
\usage{lipset}
\format{A data frame with 9 Q-sorts sorting 33 statements.}
\source{Brown, S. R., 1980. \emph{Political subjectivity: Applications of Q methodology in political science}, New Haven, CT: Yale University Press.}
\keyword{datasets}
