\name{loa.and.flags}
\alias{loa.and.flags}
\title{Q methodology: show factor loadings next to flags}
\description{Prints a table with factor loadings and flagged Q-sorts indicated with a star.}
\usage{loa.and.flags(results, nload=F, ...)}

\arguments{
  \item{results}{an object of Q method results).}
  \item{nload}{logical; print number of flagged Q-sorts).}
  \item{...}{further arguments.}
}
\details{
Simple function to explore flagged Q-sorts
}

\author{Aiora Zabala}