\name{q.fcolors}
\alias{q.fcolors}
\title{Q Methodology: assign colors to factors.}
\description{
  This function assigns colors to factors in an object of Q method results returned by \code{\link{qmethod}}.
}
\usage{
  q.fcolors(results, fcolors = NULL, color.scheme = "Set1")
}

\arguments{
  \item{results}{
    an object of class \code{QmethodRes}.
  }
  \item{fcolors}{
    A character vector of factor colors recognizable by R, in the order in which factors are extracted by \code{\link{qmethod}}.
    Used for coloring graphs and distinguishing statements.
    Defaults to \code{NULL}. in which case factors are colored automatically by \pkg{RColorBrewer}.
  }
  \item{color.scheme}{
    If \code{fcolors = NULL}, a character vector of length 1 from color schemes available in \pkg{RColorBrewer}, suitable for qualitative data (\code{type = "qual"}).
    Defaults to \code{"Set1"}.
  }
}

\details{
  Factor interpretation can be aided by visualizations, such as \code{\link{q.scoreplot}}.
  This function improves orientation by adding a consistent set of colors across all plots, where a factor can always be identified by a color.

  These colors can be manually entered using the \code{fcolors} argument, or, if left blank, colors can be automatically generated by \pkg{RColorBrewer}.
}

\value{
  Returns the object \code{results} of class \code{QmethodRes}, with the factor colors added or overwritten as a named vector under \code{results$brief$fcolors}.
  The object is returned invisibly.
}

\author{Maximilian Held}

\seealso{
  \code{\link{q.scoreplot}}
}

\examples{
  data(lipset)
  results <- qmethod(lipset[[1]], nfactors = 3, rotation = "varimax")
  results <- q.fcolors(results = results, fcolors = c("black", "red", "green"))  # for manual colors
  plot(results)  # to see colored factors
  results <- q.fcolors(results = results)  # for automatic colors
  plot(results)  # to see colored factors
}
