\name{QmethodRes-class}
\Rdversion{1.1}
\docType{class}
\alias{QmethodRes-class}

\title{Class "QmethodRes"}
\description{
An object with the full results of Q Methodology analysis. This object is automatically generated using \link{qmethod}.  Not intended to be created ex-ante.
}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Methods}{
\code{\link{plot.QmethodRes}}, \code{\link{summary.QmethodRes}}, \code{\link{print.QmethodRes}}
}

\author{
Aiora Zabala
}

\seealso{
See \code{\link{qmethod}} in this package.
}