\name{keyneson}
\docType{data}
\alias{keyneson}
\title{Keyneson, a Q Sort on Taxation and the Economy}
\description{
	A Q-methodological study on taxation and the economy, including a concourse, sample and Q-sorts.
	Go to \url{https://github.com/maxheld83/keyneson} for more information.
}
\usage{keyneson}
\format{
	A directory in the package root, including subdirectories as documented in the package documentation, and on the package \href{https://github.com/aiorazabala/qmethod/wiki/Data-management}{wiki}.
}
\source{Maximilian Held. (2015). keyneson: CiviCon-1. ZENODO. 10.5281/zenodo.14295}
\keyword{datasets}
