\name{lipset}
\docType{data}
\alias{lipset}
\title{\emph{Lipset} (1963) Q methodology dataset}
\description{Dataset about \emph{The Value Patterns of Democracy} based on Lipset (1963) to illustrate the \pkg{qmethod} package.}
\usage{lipset}
\format{A list with two objects. A data frame with 9 Q sorts sorting 33 statements and a data frame with the text corresponding to the statements.}
\source{Brown, S. R., 1980. \emph{Political subjectivity: Applications of Q methodology in political science}, New Haven, CT: Yale University Press.

Lipset, S. M., 1963. The value patterns of democracy: A case study in comparative analysis. \emph{American Sociological Review}, 28, 515-531.
}
\keyword{datasets}
