% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q.loaplot.R
\name{q.loaplot}
\alias{q.loaplot}
\title{Q methodology: create Q factor loadings plot}
\usage{
q.loaplot(results, names = TRUE, points = FALSE, alpha = 1,
  density = FALSE, grid = TRUE, rug = FALSE, red = TRUE,
  quietly = FALSE, maxvals = FALSE)
}
\arguments{
\item{results}{A results object from \code{\link{qmethod}} or a loadings table.}
}
\value{
Returns a list of lists of \pkg{ggplot2} objects, one for each pair of extracted factors.
}
\description{
Plots the loadings of Q-sorts on all factors against all factors.
}
\note{
The returned object includes all possible \emph{permutations} of factor pairs.
Because the plot of factor A against factor B is the same as factor B against factor A, only with transposed axes, half of these plots are redundant.
They are included to allow researchers to conveniently access any given pair of factors.

The function also returns a grid of all \emph{non-redundant} factor pairs, that is, the \emph{combinations} of factor pairs.

If there are factor names from \code{\link{q.fnames}} or colors from \code{\link{q.fcolors}} in \code{results}, axes will be named and colored accordingly.
Remember that factor names and colors may be specific to a particular rotation procedure, and may not be appropriate before rotations are completed.
}
\examples{
data(lipset)  # use data included with this package
results <- qmethod(lipset[[1]], nfactors=3, rotation="varimax")  # run analysis
loaplots <- q.loaplot(results = results)  # produces plots list
loaplots$f1$f2  # retrieves loaplot of f1 vs f2
}
\author{
Maximilian Held
}
